添加了球体和球壳的Shape 质点也可以指定 Shape，可用于辅助渲染

+++ src/model/physics/rigid/particle.cpp
@@ -10,7 +10,7 @@ namespace egret
 {
     void Particle::applyForce(const double time)
     {
-        const Eigen::Vector3f acceleration{getJoinForce().force / m_mass};
+        const Eigen::Vector3d acceleration{getJoinForce().force / m_mass};
         m_position += m_speed * time + 0.5 * acceleration * time * time;
         m_speed += acceleration * time;
     }
@@ -23,10 +23,10 @@ namespace egret
     void Particle::rotateMandatory(const Axis& axis, double radians)
     {
         // 1. 将旋转轴方向归一化
-        Eigen::Vector3f u = axis.rotationAxis.normalized();
+        Eigen::Vector3d u = axis.rotationAxis.normalized();
 
         // 2. 将旋转平移到原点（将点相对于轴上的基点表示）
-        const Eigen::Vector3f p_rel = m_position - axis.basePoint;
+        const Eigen::Vector3d p_rel = m_position - axis.basePoint;
 
         // 3. 构建旋转矩阵（罗德里格斯旋转公式）
         // R = I + sin(theta)*[u]_× + (1-cos(theta))*[u]_×^2
@@ -35,18 +35,18 @@ namespace egret
         const double sin_theta = std::sin(radians);
 
         // 叉积矩阵 [u]_×
-        Eigen::Matrix3f cross_matrix;
+        Eigen::Matrix3d cross_matrix;
         cross_matrix << 0, -u(2), u(1),
             u(2), 0, -u(0),
             -u(1), u(0), 0;
 
         // 旋转矩阵
-        const Eigen::Matrix3f rotation_matrix = Eigen::Matrix3f::Identity()
+        const Eigen::Matrix3d rotation_matrix = Eigen::Matrix3d::Identity()
             + sin_theta * cross_matrix
             + (1 - cos_theta) * (cross_matrix * cross_matrix);
 
         // 4. 应用旋转
-        const Eigen::Vector3f p_rotated = rotation_matrix * p_rel;
+        const Eigen::Vector3d p_rotated = rotation_matrix * p_rel;
 
         // 5. 平移回原位
         m_position = p_rotated + axis.basePoint;
@@ -58,7 +58,7 @@ namespace egret
         return pow(distance, 2) * m_mass;
     }
 
-    Eigen::Vector3f Particle::getTorque(const Eigen::Vector3f& base)
+    Eigen::Vector3d Particle::getTorque(const Eigen::Vector3d& base)
     {
         return (m_position - base).cross(getJoinForce().force);
     }
@@ -72,12 +72,12 @@ namespace egret
         return joinForce;
     }
 
-    Eigen::Vector3f Particle::getMomentum()
+    Eigen::Vector3d Particle::getMomentum()
     {
         return m_mass * m_speed;
     }
 
-    Eigen::Vector3f Particle::getAngularMomentum(const Eigen::Vector3f& base)
+    Eigen::Vector3d Particle::getAngularMomentum(const Eigen::Vector3d& base)
     {
         return (m_position - base).cross(getMomentum());
     }
