+++ src/model/physics/rigid/physical_entity.cpp
@@ -10,12 +10,12 @@ namespace egret
     {
     }
 
-    PhysicalEntity::PhysicalEntity(Eigen::Vector3f position, Eigen::Vector3f speed, const double mass):
+    PhysicalEntity::PhysicalEntity(Eigen::Vector3d position, Eigen::Vector3d speed, const double mass):
         m_position(std::move(position)), m_speed(std::move(speed)), m_mass(mass)
     {
     }
 
-    PhysicalEntity::PhysicalEntity(const Eigen::Vector3f& position, const Eigen::Vector3f& speed, const double mass,
+    PhysicalEntity::PhysicalEntity(const Eigen::Vector3d& position, const Eigen::Vector3d& speed, const double mass,
         std::vector<Force> forces): PhysicalEntity(position, speed, mass)
     {
         m_forces = std::move(forces);
